requirements.txt
django>=6.0,<7.0
google-generativeai>=0.8.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0


